Revert "Merge remote-tracking branch 'MailCore/master'"
This reverts commit fd251857c52386574e1349ab081ccdc93b60b5d5, reversing
changes made to a2b9d232fd470786813990ce49c375029271f2ce.

+++ src/core/basetypes/icu-ucsdet/ustring.cpp
@@ -997,7 +997,7 @@ u_strlen(const UChar *s)
     while(*t != 0) {
       ++t;
     }
-    return (int32_t)(t - s);
+    return t - s;
 #endif
 }
